Refactor: Setup data and domain layer for top anime
This commit introduces the necessary components for fetching top anime data from the API.

Key changes:
- Added `TopAnimeRequest` data class and `QueryMapping` interface for handling API request parameters.
- Created `PagingRequest` abstract class to handle common pagination parameters.
- Introduced `ParamsEmptyException` for cases where use case parameters are missing.
- Implemented `TopAnime` model, including adding `pagination` and `PageItem` to `ApiResponse`.
- Added `TopService` interface with a method for fetching top anime.
- Created `TopRepository` and `TopRepositoryImpl` for data handling.
- Implemented `GetTopAnimeUseCase` to provide the business logic for fetching top anime.
- Added `RepositoryModule` to the DI graph for providing the `TopRepository`.
- Modified `BaseRemoteDataSource` to accept responses with a non-200 status code if `data` is not null.
- Updated `NavigationExtension` with a new `setupWithNavController2` function for `BottomNavigationView` to improve fragment handling.
- Changed `screenState` in `BaseViewModel` to be a `MutableSharedFlow` for better event handling.
- Modified `observeLatest` extension function to accept an optional `lifecycleState` parameter.
- Added data binding to `fragment_home.xml`.
- Updated `HomeFragment` to bind the `HomeViewModel` to the layout.
- Created `HomeViewModel` with a placeholder `getTopAnime` function.
- Added the `SeasonsService` and `TopService` to the `NetworkModule`.

+++ app/src/main/java/com/lambao/mrbeast/data/model/top_anime/TopAnime.kt
@@ -31,7 +31,7 @@ data class TopAnime(
     @Expose @SerializedName("aired") val aired: Aired? = null,
     @Expose @SerializedName("duration") val duration: String? = null,
     @Expose @SerializedName("rating") val rating: String? = null,
-    @Expose @SerializedName("score") val score: Int? = null,
+    @Expose @SerializedName("score") val score: Double? = null,
     @Expose @SerializedName("scored_by") val scoredBy: Int? = null,
     @Expose @SerializedName("rank") val rank: Int? = null,
     @Expose @SerializedName("popularity") val popularity: Int? = null,
